# resources/api_doc_template.yaml
swagger: "2.0"
info:
  description: "This API documentation details the resources for the Code Stroke Backend."
  version: "$VERSION"
  title: "Code Stroke Backend"
host: "$HOST"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /cases/view/:
    get:
      tags:
      - "Cases"
      summary: "View a list of cases."
      description: "Returns a list of all cases in the database."
      operationId: "getCases"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              success: 
                type: boolean
                default: True
              result:
                type: array
                items:
                  $ref: "#/definitions/CASES"
        401:
          description: "Authentication failed."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
      security:
        - requiresClinician: []
  /cases/add/:
    post:
      tags:
        - "Cases"
      summary: "Add a new case."
      description: "Add a new case to the database."
      operationId: "addCase"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: True
          description: "Case details."
          schema:
            $ref: "#/definitions/CASES"
      responses:
        200: 
          description: "OK"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                default: True
              case_id:
                type: "integer"
        400: 
          description: "Request data was invalid."
          schema: 
            type: "object"
            properties:
              success: 
                type: "boolean"
                default: False
              error_type:
                  type: "string"
                  default: "request"
        500:
          description: "Unspecified error. Please contact the developing team."
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                default: False
      security:
      - requiresClinician: []
  /delete/{case_id}/:
    delete:
      tags:
        - "Case Operations"
      summary: "Delete a case."
      operationId: "deleteCase"
      description: "Permanently deletes a case from the database."
      security:
        - requiresClinician: []
      parameters:
        - name: "case_id"
          in: "path"
          description: "ID of case to delete."
          required: true
          type: "integer"
      responses:
        200: 
          description: "OK"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                default: True
        401:
          description: "Authentication failed."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
  /acknowledge/{case_id}/:
    post:
      tags:
        - "Case Operations"
      summary: "Acknowledge an incoming case notification"
      description: "Acknowledges that a notification for an incoming case has been seen, and will broadcast this acknowledgement and an ETA (if location data is provided) as a new notification."
      operationId: "acknowledgeCase"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "case_id"
          required: True
          description: "The ID of the case to be acknowledged."
          type: integer
        - in: "body"
          name: "body"
          required: True
          description: "The acknowledging clinician's current location details."
          schema:
            type: "object"
            properties:
              initial_location_lat:
                type: "string"
                example: "-37.8136"
              initial_location_long:
                type: "string"
                example: "144.9631"
      responses:
        200: 
          description: "OK"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                default: True
        401:
          description: "Authentication failed."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
      security:
      - requiresClinician: []    
...
securityDefinitions:
  requiresClinician:
    type: "basic"
    description: "This is the standard 2FA method for a clinician. The time-based token based on the shared secret should be concatenated onto the password with a colon i.e. `password:token` - if using the headers directly, the whole string should be of the form `username:password:token` before being base64 encoded."
  requiresAdmin:
    type: "basic"
    description: "This route requires administrator credentials, which are just a standard username and password. These resources should not generally be called from the ordinary frontend clients and may be deprecated in the future."
definitions:
  FailedAuthClinician:
    type: "object"
    properties:
      success: 
        type: boolean
        default: False
      error_type:
        type: string
        enum: ["auth"]